part and quit done
