+++ includes/Manager.hpp
@@ -38,6 +38,7 @@ class Manager
 		static	int			topicAction( void );
 		static	int			privAction( const Clients &client, std::vector<std::string> splits );
 		static	int			inviteAction( std::string nickName, int clientId );
+		static	int			partAction(std::string channelName, int clientId, std::string partMessage);
 		static	int			runChanActions( std::vector<std::string> splits, int clientId);
 		static	void		sendIrcMessage(std::string message, int id);
 		static	void		BroadcastMessageChan(Channel &channel, std::string message);
